Problem: non-consolidated tx validation (CRO-175)
Solution: moved the abci tx validation functionality into the dedicated crate
