+++ chain-tx-validation/Cargo.toml
@@ -9,6 +9,4 @@ edition = "2018"
 [dependencies]
 chain-core = { path = "../chain-core" }
 secp256k1zkp = { git = "https://github.com/crypto-com/rust-secp256k1-zkp.git", rev = "a1e70e1f0bb65d3a97f100f79d99c85395d897f8", features = ["recovery", "endomorphism"] }
-
-[dev-dependencies]
-parity-codec = { version = "4.0.0" }
+parity-codec = { features = ["derive"], version = "4.0.0" }
